--- src/model/ca.rs
@@ -1,136 +1,0 @@
-use openssl::asn1::Asn1Time;
-use openssl::bn::{BigNum, MsbOption};
-use openssl::error::ErrorStack;
-use openssl::hash::MessageDigest;
-use openssl::pkey::{PKey, PKeyRef, Private};
-use openssl::rsa::Rsa;
-use openssl::x509::extension::{
-    AuthorityKeyIdentifier, BasicConstraints, KeyUsage, SubjectAlternativeName, SubjectKeyIdentifier,
-};
-use openssl::x509::{X509NameBuilder, X509Ref, X509Req, X509ReqBuilder, X509};
-
-pub struct CA {
-    key_pair: PKey<Private>,
-    certificate: X509,
-}
-
-impl CA {
-    pub fn new() -> Result<Self, ErrorStack> {
-        let (certificate, key_pair) = generate_ca_cert_and_key_pair()?;
-        Ok(Self { key_pair, certificate })
-    }
-
-    pub fn sign(&self, key_pair: &PKey<Private>) -> Result<X509, ErrorStack> {
-        generate_x509_cert(key_pair, &self.certificate, &self.key_pair)
-    }
-}
-
-fn generate_x509_request(key_pair: &PKey<Private>) -> Result<X509Req, ErrorStack> {
-    let mut req_builder = X509ReqBuilder::new()?;
-    req_builder.set_pubkey(key_pair)?;
-
-    let mut x509_name = X509NameBuilder::new()?;
-    x509_name.append_entry_by_text("C", "IT")?;
-    x509_name.append_entry_by_text("O", "Prima")?;
-    x509_name.append_entry_by_text("CN", "prima.localauth0.com")?;
-    let x509_name = x509_name.build();
-    req_builder.set_subject_name(&x509_name)?;
-
-    req_builder.sign(key_pair, MessageDigest::sha256())?;
-    let req = req_builder.build();
-    Ok(req)
-}
-
-fn generate_ca_cert_and_key_pair() -> Result<(X509, PKey<Private>), ErrorStack> {
-    let rsa = Rsa::generate(2048)?;
-    let key_pair = PKey::from_rsa(rsa)?;
-
-    let mut x509_name = X509NameBuilder::new()?;
-    x509_name.append_entry_by_text("C", "IT")?;
-    x509_name.append_entry_by_text("O", "Prima CA")?;
-    x509_name.append_entry_by_text("CN", "Prima CA")?;
-    let x509_name = x509_name.build();
-
-    let mut cert_builder = X509::builder()?;
-    cert_builder.set_version(2)?;
-    let serial_number = {
-        let mut serial = BigNum::new()?;
-        serial.rand(159, MsbOption::MAYBE_ZERO, false)?;
-        serial.to_asn1_integer()?
-    };
-    cert_builder.set_serial_number(&serial_number)?;
-    cert_builder.set_subject_name(&x509_name)?;
-    cert_builder.set_issuer_name(&x509_name)?;
-    cert_builder.set_pubkey(&key_pair)?;
-    let not_before = Asn1Time::days_from_now(0)?;
-    cert_builder.set_not_before(&not_before)?;
-    let not_after = Asn1Time::days_from_now(365)?;
-    cert_builder.set_not_after(&not_after)?;
-
-    cert_builder.append_extension(BasicConstraints::new().critical().ca().build()?)?;
-    cert_builder.append_extension(KeyUsage::new().critical().key_cert_sign().crl_sign().build()?)?;
-
-    let subject_key_identifier = SubjectKeyIdentifier::new().build(&cert_builder.x509v3_context(None, None))?;
-    cert_builder.append_extension(subject_key_identifier)?;
-
-    cert_builder.sign(&key_pair, MessageDigest::sha256())?;
-    let cert = cert_builder.build();
-
-    Ok((cert, key_pair))
-}
-
-fn generate_x509_cert(
-    key_pair: &PKey<Private>,
-    ca_cert: &X509Ref,
-    ca_key_pair: &PKeyRef<Private>,
-) -> Result<X509, ErrorStack> {
-    let req = generate_x509_request(key_pair)?;
-
-    let mut cert_builder = X509::builder()?;
-    cert_builder.set_version(2)?;
-    let serial_number = {
-        let mut serial = BigNum::new()?;
-        serial.rand(159, MsbOption::MAYBE_ZERO, false)?;
-        serial.to_asn1_integer()?
-    };
-    cert_builder.set_serial_number(&serial_number)?;
-    cert_builder.set_subject_name(req.subject_name())?;
-    cert_builder.set_issuer_name(ca_cert.subject_name())?;
-    cert_builder.set_pubkey(key_pair)?;
-    let not_before = Asn1Time::days_from_now(0)?;
-    cert_builder.set_not_before(&not_before)?;
-    let not_after = Asn1Time::days_from_now(365)?;
-    cert_builder.set_not_after(&not_after)?;
-
-    cert_builder.append_extension(BasicConstraints::new().build()?)?;
-
-    cert_builder.append_extension(
-        KeyUsage::new()
-            .critical()
-            .non_repudiation()
-            .digital_signature()
-            .key_encipherment()
-            .build()?,
-    )?;
-
-    let subject_key_identifier =
-        SubjectKeyIdentifier::new().build(&cert_builder.x509v3_context(Some(ca_cert), None))?;
-    cert_builder.append_extension(subject_key_identifier)?;
-
-    let auth_key_identifier = AuthorityKeyIdentifier::new()
-        .keyid(false)
-        .issuer(false)
-        .build(&cert_builder.x509v3_context(Some(ca_cert), None))?;
-    cert_builder.append_extension(auth_key_identifier)?;
-
-    let subject_alt_name = SubjectAlternativeName::new()
-        .dns("auth0-proxy")
-        .dns("auth0")
-        .build(&cert_builder.x509v3_context(Some(ca_cert), None))?;
-    cert_builder.append_extension(subject_alt_name)?;
-
-    cert_builder.sign(ca_key_pair, MessageDigest::sha256())?;
-    let cert = cert_builder.build();
-
-    Ok(cert)
-}
